Moved internals into a package of their own
